Move EpochLogger to separate file
This commit moves the `EpochLogger` class to a separate file. The reason to do this is to avoid the scoping issues we had before with Pickle then it tried to serialize the Node2Vec model, as it was unable to find `EpochLogger`.
